Simplify flow: single edge function returns 7-day signed URLs
Replaces the email + token-verification flow with a single
request-documents edge function that:
  - records the request
  - watermarks each requested PDF with the company name + date
  - returns 7-day signed URLs directly to the page

Removes the now-unused verify-documents function, retrieve.html,
verify.html, the access_tokens migration, and the magic_link
migration. Drops Resend / EMAIL_FROM / VERIFY_FUNCTION_URL from
config since no email is sent.

Frontend updates:
  - inline script now renders download cards on success
  - NDA + form-field selectors use real IDs (ndaAgreed, fullName, company)
  - escapeHtml helper added to safely render user-supplied values

DEPLOYMENT.md rewritten to match the simplified flow.
