+++ .env.example
@@ -7,30 +7,16 @@ SUPABASE_URL=https://YOUR_PROJECT_REF.supabase.co
 # Publishable key — safe to embed in client-side JS.
 SUPABASE_PUBLISHABLE_KEY=sb_publishable_...
 
-# Service role key — SERVER SIDE ONLY (Edge Functions + Vercel API routes).
+# Service role key — SERVER SIDE ONLY (Edge Functions).
 # supabase secrets set SUPABASE_SERVICE_ROLE_KEY=<value>
 SUPABASE_SERVICE_ROLE_KEY=eyJ...
 
 # ── Vercel environment variables ──────────────────────────────────────────────
-# Set all of these in Vercel Dashboard → Project → Settings → Environment Variables.
+# Set these in Vercel Dashboard → Project → Settings → Environment Variables.
 
-# Full URL of the request-documents edge function (handles form submission)
+# Full URL of the request-documents edge function.
+# This single function generates 7-day signed URLs for the requested PDFs.
 EDGE_FUNCTION_URL=https://YOUR_PROJECT_REF.supabase.co/functions/v1/request-documents
-
-# Full URL of the verify-documents edge function (validates token, returns links)
-VERIFY_FUNCTION_URL=https://YOUR_PROJECT_REF.supabase.co/functions/v1/verify-documents
-
-# Public domain of the Vercel deployment (used to build retrieval links in emails)
-SITE_BASE_URL=https://your-project.vercel.app
-
-# ── Email provider ────────────────────────────────────────────────────────────
-# Resend (https://resend.com) is the recommended provider.
-# Set as an Edge Function secret: supabase secrets set RESEND_API_KEY=re_...
-RESEND_API_KEY=re_...
-
-# From address — must be a verified sender domain in Resend/SendGrid.
-# supabase secrets set EMAIL_FROM=trust@yourdomain.com
-EMAIL_FROM=trust@yourdomain.com
 
 # ── Supabase Storage ──────────────────────────────────────────────────────────
 # Create a PRIVATE bucket named "audit-docs" in Supabase Dashboard → Storage.
@@ -43,15 +29,11 @@ EMAIL_FROM=trust@yourdomain.com
 #   dpa.pdf           → Data Processing Agreement
 #   questionnaire.pdf → Security questionnaire (CAIQ/SIG/VSAQ)
 
-# ── Deploy Edge Functions ────────────────────────────────────────────────────
+# ── Deploy Edge Function ─────────────────────────────────────────────────────
 # supabase functions deploy request-documents --no-verify-jwt
-# supabase functions deploy verify-documents  --no-verify-jwt
 #
-# Set secrets (Edge Functions need these at runtime):
+# Set secrets (Edge Function needs these at runtime):
 # supabase secrets set SUPABASE_SERVICE_ROLE_KEY=eyJ...
-# supabase secrets set SITE_BASE_URL=https://your-project.vercel.app
-# supabase secrets set RESEND_API_KEY=re_...
-# supabase secrets set EMAIL_FROM=trust@yourdomain.com
 
 # ── Run DB migrations ─────────────────────────────────────────────────────────
 # supabase db push
@@ -59,5 +41,4 @@ EMAIL_FROM=trust@yourdomain.com
 # Migrations applied (in order):
 #   20260501000000_document_requests.sql        — main requests table
 #   20260501000001_document_requests_add_links.sql — adds links jsonb column
-#   20260502000000_short_urls.sql               — short URL redirects
-#   20260506000000_access_tokens.sql            — secure access tokens for email flow
+#   20260502000000_short_urls.sql               — short URL redirects (optional)
