--- website/retrieve.html
@@ -1,214 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8" />
-  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <title>Retrieve Documents – Trust Center</title>
-  <link rel="stylesheet" href="/css/styles.css" />
-  <style>
-    .retrieve-wrap {
-      max-width: 680px;
-      margin: 80px auto;
-      padding: 0 20px;
-    }
-    .state-box {
-      background: #fff;
-      border: 1px solid #e5e7eb;
-      border-radius: 16px;
-      padding: 48px 40px;
-      text-align: center;
-      box-shadow: 0 4px 24px rgba(0,0,0,0.07);
-    }
-    .state-icon { font-size: 56px; margin-bottom: 16px; }
-    .state-box h2 { margin: 0 0 12px; font-size: 24px; color: #111827; }
-    .state-box p  { margin: 0 0 8px; color: #6b7280; font-size: 15px; line-height: 1.6; }
-
-    .doc-card {
-      background: #f9fafb;
-      border: 1px solid #e5e7eb;
-      border-radius: 12px;
-      padding: 20px 24px;
-      margin-top: 16px;
-      display: flex;
-      align-items: center;
-      justify-content: space-between;
-      gap: 16px;
-      text-align: left;
-    }
-    .doc-info { display: flex; align-items: center; gap: 14px; }
-    .doc-icon { font-size: 32px; }
-    .doc-name { font-weight: 600; color: #111827; margin: 0 0 4px; }
-    .doc-sub  { font-size: 13px; color: #6b7280; margin: 0; }
-    .btn-download {
-      background: #2563eb;
-      color: #fff;
-      padding: 10px 22px;
-      border-radius: 8px;
-      text-decoration: none;
-      font-weight: 600;
-      font-size: 14px;
-      white-space: nowrap;
-      flex-shrink: 0;
-    }
-    .btn-download:hover { background: #1d4ed8; }
-
-    .uses-badge {
-      display: inline-block;
-      background: #f0fdf4;
-      color: #16a34a;
-      border: 1px solid #bbf7d0;
-      border-radius: 20px;
-      padding: 4px 14px;
-      font-size: 13px;
-      font-weight: 500;
-      margin-top: 12px;
-    }
-    .warning-box {
-      background: #fef3c7;
-      border-left: 4px solid #f59e0b;
-      border-radius: 0 8px 8px 0;
-      padding: 12px 16px;
-      margin-top: 24px;
-      text-align: left;
-      font-size: 14px;
-      color: #92400e;
-    }
-  </style>
-</head>
-<body>
-  <header class="site-header">
-    <div class="container header-inner">
-      <a href="/" class="site-logo">Trust Center</a>
-    </div>
-  </header>
-
-  <main>
-    <div class="retrieve-wrap">
-
-      <!-- Loading state -->
-      <div id="state-loading" class="state-box">
-        <div class="state-icon">⏳</div>
-        <h2>Verifying your link…</h2>
-        <p>Please wait while we validate your access token.</p>
-      </div>
-
-      <!-- Success state (hidden until verified) -->
-      <div id="state-success" class="state-box" style="display:none;">
-        <div class="state-icon">✅</div>
-        <h2>Your documents are ready</h2>
-        <p id="success-subtitle"></p>
-        <div id="doc-list"></div>
-        <div class="warning-box">
-          🔒 These documents are governed by the NDA you accepted.
-          Do not share links or distribute files outside your organisation.
-        </div>
-        <div id="uses-remaining"></div>
-      </div>
-
-      <!-- Error state (hidden until error) -->
-      <div id="state-error" class="state-box" style="display:none;">
-        <div class="state-icon">❌</div>
-        <h2 id="error-title">Link invalid or expired</h2>
-        <p id="error-msg">This link is no longer valid.</p>
-        <a href="/#request-docs" style="display:inline-block; margin-top:24px;" class="btn btn-primary">
-          ← Request new documents
-        </a>
-      </div>
-
-    </div>
-  </main>
-
-  <footer class="site-footer">
-    <div class="container footer-inner">
-      <span>© <span id="year"></span> Trust Center</span>
-    </div>
-  </footer>
-
-  <script src="/api/config"></script>
-  <script>
-    document.getElementById('year').textContent = new Date().getFullYear();
-
-    const VERIFY_URL = window.ENV?.VERIFY_FUNCTION_URL ||
-      'https://YOUR_PROJECT_REF.supabase.co/functions/v1/verify-documents';
-
-    function show(id) {
-      ['state-loading','state-success','state-error'].forEach(s => {
-        document.getElementById(s).style.display = s === id ? '' : 'none';
-      });
-    }
-
-    function escHtml(str) {
-      return String(str).replace(/[&<>"']/g, c =>
-        ({ '&':'&amp;','<':'&lt;','>':'&gt;','"':'&quot;',"'":'&#39;' }[c]));
-    }
-
-    async function verifyToken(token) {
-      const res = await fetch(VERIFY_URL, {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ token }),
-      });
-      return { ok: res.ok, data: await res.json().catch(() => ({})) };
-    }
-
-    async function init() {
-      const params = new URLSearchParams(location.search);
-      const token = params.get('token');
-
-      if (!token) {
-        document.getElementById('error-title').textContent = 'No token provided';
-        document.getElementById('error-msg').textContent =
-          'This page requires a retrieval token from your email. Please use the link in your email.';
-        show('state-error');
-        return;
-      }
-
-      try {
-        const { ok, data } = await verifyToken(token);
-
-        if (!ok || !data.success) {
-          document.getElementById('error-title').textContent = 'Link invalid or expired';
-          document.getElementById('error-msg').textContent =
-            data.error || 'This link is no longer valid. Please submit a new request.';
-          show('state-error');
-          return;
-        }
-
-        // Render success
-        document.getElementById('success-subtitle').textContent =
-          `Prepared for ${escHtml(data.name)} · ${escHtml(data.company)}`;
-
-        const list = document.getElementById('doc-list');
-        list.innerHTML = data.documents.map(doc => `
-          <div class="doc-card">
-            <div class="doc-info">
-              <span class="doc-icon">${escHtml(doc.icon)}</span>
-              <div>
-                <p class="doc-name">${escHtml(doc.name)}</p>
-                <p class="doc-sub">PDF · Watermarked · Confidential</p>
-              </div>
-            </div>
-            <a class="btn-download" href="${escHtml(doc.url)}" target="_blank" rel="noopener noreferrer">
-              📥 Download
-            </a>
-          </div>`).join('');
-
-        if (typeof data.usesRemaining === 'number' && data.usesRemaining > 0) {
-          document.getElementById('uses-remaining').innerHTML =
-            `<span class="uses-badge">✓ ${data.usesRemaining} access${data.usesRemaining !== 1 ? 'es' : ''} remaining on this link</span>`;
-        }
-
-        show('state-success');
-
-      } catch {
-        document.getElementById('error-title').textContent = 'Network error';
-        document.getElementById('error-msg').textContent =
-          'Could not reach the server. Please check your connection and try again.';
-        show('state-error');
-      }
-    }
-
-    init();
-  </script>
-</body>
-</html>
